feat(metadata,gateway): implement request limiter

+++ services/metadata/internal/app/quota/repository/repository.go
@@ -11,7 +11,7 @@ import (
 type Quota struct {
 	gormext.UniversalModel
 	UserID              uuid.UUID
-	MaxRequestPerMinute int
+	MaxRequestPerMinute int64
 	MaxVolumeInMonth    int64
 	TotalUsedVolume     int64
 }
